requirements.txt
-e ./shoop
django-registration-redux==1.2
